MBS-4709: Fix square brackets in annotation text
As a side effect, the commit 65ad3b115cfe36 caused square brackets
encoded as HTML entities &#91; and &#93; to be decoded before wiki
formatting, thus being mistakenly interpreted as a wiki link.  The
assumption that HTML entities are equivalent to their decoded form
is not true in the particular case of square brackets.

This patch mostly reverts 65ad3b115cfe36 and subsequent commits,
thus fixing display of square brackets entered as HTML entities so
as not to create a link, and showing HTML entities being replaced
with their decoded form in annotation diff again.