Add USE_SET_DATABASE_HEADER
It turns out that the feature provided by USE_SELENIUM_HEADER would be
useful in other places, e.g. the sitemap tests. This commit renames the
header to 'mb-set-database' and the corresponding DBDefs flag to
'USE_SET_DATABASE_HEADER', allowing it to set any database during the
duration of the request.

I've modified the incremental sitemaps test to use this, so it can now
be run without needing a weird database configuration.