Update the yarn pubkey again
The yarn pubkey expired a couple weeks ago and they updated it. We
pulled the new one in 323c67e718cefc215bd5ce6377216f4d13fc9561.

However, a few days ago they changed the key again for some reason, so
we have to pull it again. There are some issue reports on the yarn repo
where they're investigating the cause of the key change.