Remove pause commands from selenium tests
`pause` has two problems:

 1. Time is wasted if the duration is too long.
 2. Tests break if the duration is too short (and the correct length is
    system-dependent).

This replaces certain `pause` commands with a generic `waitUntil`
command that waits for some condition to become true before proceeding.

In the Artist_Credit_Editor tests, the hidden input checks were
problematic because there wasn't a good hook to use with waitUntil. I
changed them to submit the form and check that the edit data is
correct, which is a better test in any case.

I also removed all uses of the `fireEvent` command: not only was it not
necessary, it caused many transient timing issues. My guess is that
firing these extra events caused additional renders by Knockout or React
which then overwrite certain input fields we were typing in.