Handle falsey link type descriptions (MBS-9984)
677ec64971a1a294094b796863fe62c1f90c6f31 broke this by unconditionally
passing linkType.description to l_relationships.

A link type can have an empty description if it's only used for
grouping. Normally these can't be selected and this code would never be
hit, but in the case of MBS-9984, there already exist 'child'
relationships from before it became a grouping type.