Run yarn in compile_resources.sh (#920)
This seems like a good idea, in case people forget to do it when
updating their git checkout or switching branches. I thought about this
after merging the Webpack PR.

If node_modules/ is up-to-date and this command is a no-op, it only
takes about ~0.5s to run on my machine.