Split field types into writable and read-only
On most forms we want the fields to be read-only, because no JavaScript
code should write to them. On others, like the work edit form or user
preferences form, the fields are meant to be "writable."

Prior to this commit, we had all properties on the *FieldT types set as
covariant. This worked because we never actually wrote to any
properties: we used the lens utility to create new, immutable copies.

There are issues with the lens utility, however, as it uses the unsafe
and deprecated existential type (`*`), and is nearly impossible to type
correctly. I'd like to replace it with a different (and much easier to
use) utility in a subsequent commit; however, this relies on us having
the *FieldT types able to be written to in some cases.

I removed FieldRoleT and duplicated all of the properties to every type.
Unfortunately, type spreads are buggy and break variance checks in some
cases. There are other issues with object intersections, too.