Safer createField/pushField
createField was written in such a way that you could pass any structure
of objects/arrays/primitives, and it'd convert it into the correct
CompoundFieldT/RepeatableFieldT/FieldT hierarchy, recursively.

This wasn't possible to write correct or safe types for, and is also a
lot more general than we need it to be. There's no case where we need to
create a structure recursively. So createField now just creates a single
FieldT from a single value. A separate createCompoundField utility has
been added to create a CompoundFieldT.

The new types are safer since they don't have to deal with a recursive
structure that Flow can't check.