MBS-10072: JSON WS: hyphenate begin_area, end_area
As noted in the ticket, SOLR already uses 'begin-area' and 'end-area',
and this difference is probably causing more harm than fixing it will
(see the linked Picard ticket).

The solution here is to output all four properties for an amount of
time, announce the change, and remove the old ones in the future.