MBS-10067: Convert releases-list.tt to React
This changes the TT behaviour for instruments, since it now will
have an instruments credit column on instrument pages even if
no credits exist for that particular instrument. That adds
consistency to the tables: the same number of columns will be
there at all times, while earlier the credit column would appear
and disappear depending on whether the instruments on a specific
page of the paginated results had any credits.