MBS-10065: Convert recordings-list.tt to React
This changes the TT behaviour for instruments, since it now will
have an instruments credit column on instrument pages even if
no credits exist for that particular instrument. That adds
consistency to the tables: the same number of columns will be
there at all times, while earlier the credit column would appear
and disappear depending on whether the instruments on a specific
page of the paginated results had any credits.

This also gets rid of the alternative table if recording.gid
does not exist, in favour of just special-casing the artist credit
and length columns. The displayed result is the same. This is
only needed for merges, as per commit
a54f618af949d41b86192c2e86adad1bf6a68f1e