Make *TypeT types non-opaque, unbreak AnyReactElem
These were opaque because they were all simply aliases for OptionTreeT,
but now they're not. And being opaque prevents them from being used in
disjoint union refinements (by entityType).

You'll note that this commit contains what appear to be unrelated i18n
changes. This is because removing 'export' from the *TypeT lines
revealed some previously hidden Flow errors (see issue 4466 in the Flow
project repo).

The main source of errors was

  declare type AnyReactElem = React.Element<any>;

in root/types.js, which was broken due to React not being imported. I
replaced it with the built-in React$Element, and fixing it revealed
actual type errors wherever AnyReactElem was used (I assume the type
checker previously treated it as `any`).