MBS-10088: Correctly check for cardinality on l_x_x rels
The current query used "${source_id}_cardinality = 0" if $use_cardinality; to exclude relationships affected by cardinality.

That works great, except on l_x_x relationships where *both* sides of the relationship need to be checked for cardinality. In that case, only entity1 (the final value of $source_id) was being checked, meaning if entity_1 had cardinality=1 it would always return no results, even if the entity we were looking for was entity0 with cardinality=0.

This changes it so that the condition strings are actually fully built inside the $type checks, which ensures we'll get "WHERE (entity0 IN (?) AND entity0_cardinality = 0) OR (entity1 IN (?) AND entity1_cardinality = 0)" rather than "WHERE (entity0 IN (?) OR entity1 IN (?)) AND entity1_cardinality = 0".