Remove 'unaccented' name from link attribute types
This was used, I presume, to allow lazy searching of accented instrument
names in English when the multiselect component was still used for
instruments. However, we don't use the knockout multiselect for
instruments anymore, ever since they were converted to entities years
ago. We now use the standard Autocomplete (and SOLR should handle
diacritics just fine).

Attributes that still *do* use the knockout multiselect include "vocal"
and "subject," but neither of those contain any values with diacritics
in English. Even if they did and this becomes a real problem, the
feature is better implemented in JavaScript.