Remove build_attr_info
This was used to build a JSON tree of every link attribute type for the
relationship editor.

Other than serializing children, this duplicates the TO_JSON method
present in Entity::LinkAttributeType. We don't want multiple routines
or formats for serializing the same data.

Instead of passing the result of Data::LinkAttributeType::get_tree to
build_attr_info, we now call get_all and build the children on the
client.