MBS-9413: Fix for old 'remove relationship' edits
Older 'Remove Relationship' edits seem to only have root_id/root_name
for attributes, which we can use for the id/name and assume the type is
its own root.

Another issue is that attribute gids aren't present in these edits. The
LinkAttributesArray type allows these to be optional, but not undef.