Fix DBDefs import in CritiqueBrainzReview
'DBDefs' can't be imported in scripts that run in the browser. It can
contain sensitive configuration values, so is specifically excluded from
the build by Webpack.

'DBDefs-client' must be used instead, as it's available in the browser
and specifically includes only the %CLIENT_DEFS listed in
script/dbdefs_to_js.pl.

Resolves
https://sentry.metabrainz.org/metabrainz/musicbrainz-server/issues/52114/