MBS-9028: Skip redirects when loading Wikipedia extracts
Loading redirects as extracts is generally unwanted, since in most cases
they end up returning the extract for a more general page that isn't really
a 1:1 mapping to the entity on our side. This is especially unwanted
if it means an English redirect is being preferred over an actual proper
mapping in a different language.

This regenerates the lwp sessions for the test, since the mechanism has
changed. In a few cases the old examples were no longer valid since more
data has been added on the Wikipedia side, so they were substituted by
working examples.