Fix building test-database docker image
`test-database` docker image is currently based on `postgres:9.5.6`
which, in turn, was based on `debian:jessie`. The recent removal of
`jessie-updates` from mirrors broke using `apt-get update`, see:
https://lists.debian.org/debian-devel-announce/2019/03/msg00006.html

This patch works around this by hot-fixing `/etc/apt/sources.list`.

Note: Updating to the latest postgres:9.5 would solve this as well,
but it better has be done just before we plan to update production.