MBS-7480: Add filesize columns for cover art
This allows for storing file sizes for cover images and their
thumbnails. This'll allow us to better detect when a thumbnail isn’t
available, and also allow us to display file sizes to the user before
they download an image.

MBS hasn't been modified to actually use these fields here (they'll have
to be populated first to be useful).