Fix display of attributes on ordered rels
Normally, we display attributes (like, this) after a relationship if
they are "extra" (as in, they don't appear in the link phrase). For
ordered link types, we remove attributes from the link phrase (so that
the relationships stay grouped together under a single phrase), so that
logic doesn't work. For ordered relationships, we now display all
attributes in parentheses after the relationship.