Fix for updating track ACs with the release AC (#1025)
This is a hack, since references to React component instances are
not supposed to be saved and then be passed to setState call from
outside the component, though it technically works.

But it fixes an issue where the track artist credits would not
appear to be updated when the matching release artist credit was
changed. It achieves this by setting the component state directly,
similar to how we do inside `doneCallback` in bindingHandlers.js.

The non-hacky way is to have the release editor converted to React.