Remove artistCreditFromArray
This doesn't really do anything, since artist credits are already
arrays. (When we used Immutable.js the function existed to convert the
input to an immutable array, but that's no longer the case.)

The one thing it does additionally is copy the artist name to the
credited name if that's blank, but from what I recall that was only
useful for seeded data in the release editor.

The usage of this function would also become very awkward, as I'll be
changing the JSON format of artist credits from an array to an object in
a subsequent commit.