Remove mbs_constants.sh
Our docker images expect this to be copied to /etc/, but that's not true
in the tests image. We could copy it there in the circleci config, but
I'd prefer not making our setup more complex for two variables which
will never change. Copying the (small) declarations to each file removes
an annoying dependency and build step.