MBS-9381: Show private collections you have access to
This used to only show the number of private collections
an entity is in, and it only checked collections the
editor had access to, effectively meaning the user could
see in how many of their own private collections the entity
was in, but not which (and no other private collection
info). This fixes it so that collections that are visible
to the user (including private collections they're
collaborating on) are shown and linked, and a number
of collections the user has no access to is listed.