Show the right number of collections on sidebar
The sidebar was only showing the number of collections
the user had access to, which seems unintended.
This changes it so that the actual number of collections
will be shown.

Since there's no need to actually pass the collections
to Node, changed it to just pass the number of collections
as an integer, which is all the site needs to know.