Set GIT_* env vars in the webservice image
See docker/templates/DBDefs.pm.ctmpl

These env vars will set the corresponding configuration values in the
above template, which will prevent the server from falling back to the
default implementation, which is invoking ./script/git_info. As there
is no git client installed in the webservice image (nor .git directory,
so it wouldn't help anyway), that default implementation always fails.

The GIT_* DBDefs values are accessed, for instance, when posting an
error to Sentry.

Setting env vars in the image is exactly what the website container
already does, so this omission was likely an oversight.