Bump Flow to 0.98.1 (#1076)
Updates flow-typed/npm/redux_v3.x.x.js to remove the deprecated $Subtype
utility (though they replaced it with * which is also
deprecated, though Flow doesn't complain about it, so...).

Fixes an issue where Flow now expects a $ReadOnlyArray from
PhraseVarArgs.get. Not sure which bit in the changelog correlates to
that, but perhaps "mixed refined to an array produces a read-only
array" now.