Disable consul-template dedup mode
936d874130589b5f000a0286fba4eb441fb9260f somehow broke dedup mode on our
beta servers. This may be related to the use of `env` in the template
(the consul-template README has a vague warning about this), despite our
setting of different deduplicate prefixes in the configs on each server.
(For reference, the consul-template logs show that the process is stuck
on "listing data for template hash.")

Since it's unclear how to fix this and we only have a small number of
containers, disabling dedup mode is the simplest course of action.