MBS-9512: Reject empty recording name edits
The issue could be reproduced by simply enabling "Copy all track titles
to associated recordings," and entering spaces for one of the track
names. The spaces prevented it from triggering the "You must enter a
title and select an artist for every track" error message.

That's a contrived situation that someone may not do intentionally, but
it's easy to see how it can happen if a script goes haywire.

The validation in the release editor has been fixed to see tracks with
whitespace-only names as being empty. Further, it now checks if the name
is blank before preparing a recording edit for the server.

On the server side, /ws/js/edit has been fixed to reject such edits with
an 'empty name' error, so that they never reach the edit class.