MBS-10272: Convert the header / navbar to Bootstrap
How Bootstrap is set up:

We import the custom Bootstrap CSS from design-system, and that's it.

For the actual components, I chose *not* to use react-bootstrap for a
number of reasons:

 * It does weird things with context, causing different markup to be
   rendered on the client and server when hydrating.
 * Its components obscure what the HTML will actually look like, making
   the official Bootstrap docs hard to use.
 * It outputs large component trees (in terms of component hierarchy,
   not markup), which has performance implications for server-rendered
   pages.
 * It doesn't skip pointless client-only work (like creating event
   handlers) on the server, which also degrades performance.

A rough performance comparison of a react-bootstrap based navbar vs. a
plain React one (both passed to ReactDOMServer.renderToString) showed
the react-bootstrap one to be 5X slower. Even though it was a contrived
test for a specific component rather than a whole page, we do plan to
make heavy use of Bootstrap on our pages, and I don't think making our
pages slower on the server is a good idea if we can avoid it.

The only issue with avoiding react-bootstrap was a lack of JS for
interactive components like menus. Fortunately, basic implementations
that are "good enough" for our purposes are easy to write in React: I
added root/static/scripts/common/hooks/useDropdown.js for the menus
here.