MBS-10222: Show legal name aliases of legal name artist
An artist linked as "Performs as" can have an artist name that differs
from the person's full legal name. This code makes it so that
artist's legal name aliases are displayed after the artist link,
hopefully ensuring the full legal name is available under the
"Legal name" section of artists linked to it with the "legal name"
relationship.