MBS-10223: Use relationship credits in ArtistRoles
This makes it so that if a relationship used in ArtistRoles
(work writers or event performers) has a relationship credit
for the artist, the credit is also displayed in the ArtistRoles
display.

The search results for works don't include credits
for work relationships, so in search results the credits
will be missing until SEARCH-585 is fixed.