MBS-10235: Drop format when reusing mediums
The medium format being kept when reusing a medium causes
a lot of problems (when editors forget to unset it) yet it rarely is of any
use: a tracklist being the same doesn't imply the format will be and
in fact it's often the opposite, with the same release having
CD, vinyl and digital versions, for example. This drops the format
when reusing a medium, so that the user will have to either select
a format or select the "I don't know" checkbox for it.