MBS-1870 / MBS-8477 / MBS-10341: User tag page improvements
This converts the user tag page to React (MBS-10341).

MBS-1870: This implements pagination on the user tag page,
in the same way as it is implemented on the general page for a tag:
only the 10 first tags will be shown on the main (all-entity) page
user/foo/tag/bar, with all others being pushed onto separate
per-entity pages.

MBS-8477: By default, only tags the editor has *up*voted are shown
on the user/foo/tag/bar pages. To see tags the editor has downvoted,
a parameter 'show_downvoted' is used. The texts for tags not being used
will change depending on whether up or downvotes are being shown.