MBS-10137: Fix guess feat. on recording pages
The guessFeat function expects an entity's artistCredit to be an
observable, which was previously only true on release pages. This is now
made true on the other entity forms.

We also need a way to update the ArtistCreditEditor component when the
observable changes. This is done by storing the component instance on
the table node, and later calling setState on it. (Unfortunately, we
can't just render FormRowArtistCredit again using a different key,
because ArtistCreditEditor does some extremely bad things with
ArtistCreditBubble that leave behind stale state. These components
should hopefully be rewritten and changed to be fully controlled in the
future.)