MBS-10263: Don't insert annotation if edit is open
The annotation edit code was probably written before we could make
all edits votable. Even if it's made votable, an annotation edit will
immediately apply, and cancelling it does nothing.
This changes that so that the annotation is only changed on edit
application, so that cancelling it actually has an effect.
To avoid overwriting changes made in the meantime, the edit is rejected
if the latest annotation for the entity was created after the edit
was entered.

Votable annotation edits will not be applied at all while the vote is
open. In most cases annotations are an autoedit anyway, unless the user
explicitly asks to make it open for a vote, so in that case it follows
they aren't sure whether the annotation should look like that or not and
it is better not to show it until it's accepted.