MBS-3112: Switch order of release editor buttons
Other menus (e.g. merge menus) have Enter first, Cancel second. This
doesn't matter too much for normal clicking but it makes a lot of
difference in the tab order (when navigating with the keyboard).

In the release editor, we generally want to submit the edits if we're
in the edit note page, and otherwise navigate either forwards
or backwards, much more often
than we want to cancel the whole thing (which in my experience is
easier done by closing the window anyway).