MBS-10278: Check no collaboration/member rels exist
As far as I can tell, this currently checked each relationship to see
whether it is one of the disallowed ones for the report, meaning a
false positive will be given if there is at least one relationship for
the artist that is *not* collaboration or member of band, even if
there's also collaboration or member of band relationships.
This changes the query to check that the artist has no collaboration
nor member of band relationships as a whole, rather than checking
each relationship separately.