MBS-9981: Exclude more reltypes from PossibleCollaborations
In cases such as "Some Dude & His Orchestra", the correct
relationship is either "conductor position" or "founder" (or both!), yet
the current implementation wouldn't consider that sufficient
because it's not "member" nor "collaborator".
This changes the query to support both of those, and also "subgroup"
(which is kinda-sorta "collaborator" for bands).