More dump-entities-sql options
Most of this data isn't relevant when I need to dump data for a test,
and it can make the .sql files quite large. Better to disable these by
default and allow the user invoking the script to specify what they
need.

`$dump_types` in particular must remain disabled when generating test
data, because the types are already included in t/sql/initial.sql.

The flags are left as they were (enabled) for the SampleDataDump.