Fix release-recording merge display (MBS-10279)
Prior to this commit, there were three different subroutines for
calculating recording merges for $EDIT_RELEASE_MERGE edits. As you might
expect, these did not share a common implementation.

 1. Data::Release::determine_recording_merges:

    This was used for detecting "bad" recording merges where the artists
    differed (see root/release/merge.tt).

 2. Edit::Utils::calculate_recording_merges:

    This was added for storing `recording_merges` in the edit data, for
    display and historical purposes (see 0e41a26), but still not for
    deciding what would actually get merged. (And after 0e41a26,
    $EDIT_RELEASE_MERGE edits did not display any recording merges that
    would actually happen, only those that were predicted to happen
    when the edit was opened. Thus, the edit might say "All recordings
    for these releases are already merged" while open, yet that may
    have been false.)

 3. Data::Medium::merge:

    This is what implemented the actual merge logic. It received a
    couple fixes over the years in dae5b54 and c1992b2, both of which
    could cause the result of the edit to disagree with the display
    (assuming, at a minimum, the tracklist didn't change in the
    meantime).

These three methods have been unified into
Data::Release::determine_recording_merges, but the implementation is
largely based on the previous Data::Medium::merge method. The reasoning
for its being moved to Data::Release (and Data::Medium::merge being
removed) is that (1) a recording could appear multiple times on the
release, which needs to be factored into the merge calculation, and (2)
it doesn't make sense to merge two arbitrary mediums outside the context
of a release merge.

This commit also fixes the display issue mentioned in point 2 above. We
now display the recording merges that will actually happen when the edit
is applied, not those that were (incorrectly) calculated when the edit
was opened. Additionally, when the edit is accepted, we update the
edit_data table to save the applied recording merge info for historical
display purposes.