Reject recording merge cycles (MBS-10188)
We previously addressed this in c1992b2 by silently ignoring such
situations (but allowing the release merge to go through). It turns out
that the fix there only works for some types of merge cycles, but not
all. It also seems unsafe to allow releases with potentially misordered
recordings to be merged at all, so this is now a hard error.

Note: I added ModBot to t/sql/initial.sql, since a lot of tests use
them. ModBot has an editor id of 4, which conflicts with unrelated
tests. I had to fix the editor ids in those tests to not conflict
with ModBot.