MBS-3920: Add more matching to Featuring reports
This extends the Featuring reports with more options that also
mostly mean the artist needs to be fixed: "f/", "featuring", "ft.", "w/"
and "duet with". It does not add other similar options (such as
"con", "avec", "with") because they seem to cause too many false
positives (mostly but not only related to classical music).