MBS-10500: Allow localizing ModBot edit notes
ModBot's notes are currently always stored and displayed in English.
With this commit, a note can be stored in the database in a special
format:

  localize:{"message": "...", "args": {...}}

where "message" is a message in our gettext catalog, and "args" are any
substitution arguments needed.

When loading a note from ModBot, we check if it's in this format before
displaying it, and parse it to form a localized string using those
parameters.

This syntax can only be used in notes left by ModBot (a regular user
could not).