MBS-9111: Stricter validation for empty edit notes
We keep having issues with people adding edit notes with no content.
This blocks edit notes with only whitespace and (common) punctuation
combinations, and asks the user to provide an edit note
that will be useful to fellow editors.
It also blocks one-character ASCII notes, but not one-character notes
in wider Unicode, since glyph languages (such as Japanese) might have
a full meaningful word that is one character long.