MBS-4326: Show DiscID edits in RG and artist history
DiscID edits seem relevant enough to both the release group
the release is in and the artists in release and release group to
be displayed in those's edit histories as well.
Remove Disc ID was already being shown in the relevant places:
this changes Add and Move to do the same.