MBS-6532: Show work edits in recording / release history
Most work edits are being made based on data from releases, so
it seems reasonable for those edits to be shown on the release
edit list. Since the connection between release and work is
through a recording, it makes sense to show the edits on the
recording list as well.