MBS-4478: Improve messages for edits that create entities
When editing a medium and using a new recording, the current tooltip
claims that the entity will be created once the edit is entered (both in
the preview and... once the edit is entered). This is not really true:
it is created once the edit is *applied*. While adding a medium, though,
it really is created once the edit is entered. I changed the tooltip
to be more general so it applies in both cases.
Additionally, it doesn't really make a difference when a recording
in the new column has been removed or is being created, but we do
have that information (the former has a recording_id but the latter
does not), so I'm changing the code to actually check what is the case.
There's a third issue where the "will be created" message is never
replaced by a link to the actual recording (MBS-7814), but that's a
less straightforward fix, so let's start with this.