MBS-8328: Limit work writers to writing rels
Work writers are currently calculated as "every single artist-work rel".
This worked when we came up with it, but doesn't anymore, since
we've added several relationships that aren't really connected
to writing (such as "publisher", "commissioned by" or "dedicated to").
These shouldn't appear on the writers list, so I've hardcoded the
relationships that should be considered writing-related in Constants.