MBS-4644: Show releases that have CAA images in lists
For the artist, label and release group release lists, it is useful to
know which releases have cover art (both if you want to find which
ones to improve and if you just want to look at cover art). This adds
an icon before the release name, similarly to the video recording icon.