MBS-5049: Link to edit note syntax from edit pages
We currently only show some info about edit note syntax when entering
an edit, but not when voting on it. The info is equally useful there.
While at it, added a reminder to follow the CoC while voting:
if we're going to have text, we might as well remind people to be nice!