MBS-4826: Show RGs for recordings in artist list
A popular artist's recording list is usually a lot of the same name with
only a length to differentiate them. This makes it so that the release
groups each recording appears in are also shown, making it easier
to figure out if they're different recordings (e.g. album vs live
bootleg) or whether maybe they should be looked into for merging.