Add View::WS
The web service currently shares a Catalyst view with the website, which
can attempt to contact the template renderer if an error occurs (and
time out, since there's no renderer process in the WS container). This
solves that issue by creating a separate view for WS requests.

This has the effect that ETag and Expires headers (the latter for
slaves) are only set for WS responses, and Cache-Control is only set for
website responses.