MBS-10336: Clarify disc#.track# cell in recording page
It's not always clear what, say, "12.3" on a recording page means.
Is it disc 12, track 3 or disc 3, track 12? This adds a hover text to
at least clarify it there. If we can come up with a clearer way that
does not induce any confusion, we could just change the display and
remove the hover again.