MBS-10280: Cancel open edits from newest to oldest (#1142)
When deleting a user, we cancel all their open edits. This can cause
issues if, for example, an "Add Release Group" edit is cancelled
before its associated "Add Release" edit is (the cancelling
the RG edit will fail), or if an "Add Release"  edit is cancelled before
edits associated with that release's recordings (the recordings
won't get autoremoved and will remain orphaned).
This change makes sure we always delete the edits in reverse
chronological order, so that no edits depend on already cancelled ones.