Limit # of cache entries per call
Frequently-accessed entities will almost always be in the cache. So if
_create_cache_entries is called with thousands of objects, we can be
sure that almost none of them are accessed very often. Spending time
adding thousands of infrequently-accessed entities to the cache is
probably not a good use of time. This commit limits the number of
entries that _create_cache_entries will return to 500. This can
hopefully improve the latency of massive WS responses slightly.