MBS-5083: Display recording artist on release view
There's currently no good way of telling if the  recording artist
is different from the track artist from a release page without actually
opening the recording page for each recording.
This patch makes it so that if the track and recording ACs differ,
the recording AC is displayed under the track name (since the track
artist column, which might seem like a more obvious choice,
doesn't appear at all if all track artists match the release artist).
This is particularly useful in two cases:
a) when a user has fixed a track artist but forgotten to fix the
recording artist, or vice versa.
b) for classical music, where the recording artist should usually be different
from the track artist.

Since non-logged in visitors might be confused by the track / recording
dichotomy while looking at a tracklist and this is meant mostly for editors,
the info is only shown when logged in.