MBS-5225: Show tracklists while importing TOC from stub
When importing a discID normally, you get given a nice table
with a lot of data, including tracklist (see cdtoc/attach_list.tt).
When importing it from a CDStub, a list of "Possible Mediums" was given,
with only release title, artist, and medium number, forcing users
to choose semi-blindly.
This patch changes the table to be more similar to the one in
attach_list, with both more info and collapsible tracklists.