Remove gid sorting from WS lists
This is unnecessary, because we can rely on the order we got from the DB,
where the original query will have already had an `ORDER BY` clause.
(And that should be stable, otherwise limit/offset wouldn't work
consistently and we'd have a bug.)

This makes browsing using limit and offset more *visually* logical,
since changing the offset will now move forward in the original list in
its logical order. (This only isn't very noticeable because the offset
is typically incremented by the full limit size, changing the entire
list.)