MBS-10361: Move edit note help for better tabbing
Users will want to tab directly to the submit button on edit pages,
which the current position of the edit note help prevents (at least
in Chrome). This moves it above the note field, but makes it .small
to ensure the user doesn't confuse it with an existing edit note.